chore: move around tests and methods
